Integrate MCP functionality into existing Agave modules
- Integrate MCP schedule types (ProposerSchedule, RelaySchedule, McpSchedule)
  into leader_schedule.rs, leader_schedule_cache.rs, leader_schedule_utils.rs

- Integrate MCP fee payer validation (McpFeePayerError, SlotFeePayerTracker,
  McpFeeBreakdown, FeePhaseResult) into account_loader.rs

- Integrate MCP relay processing (ProposerShredMessage, ValidatedShred,
  SlotShredTracker, RelayShredProcessor) into sigverify_shreds.rs

- Integrate MCP proposer distribution (ProposerShred, ProposerShredBatch,
  RelayShredMessage, ProposerDistributor) into broadcast_stage.rs

- Remove standalone MCP modules from core/ (attestation_service, bankless,
  consensus_broadcast, replay, voting)

- Keep canonical MCP files: ledger/src/mcp.rs (constants) and
  ledger/src/mcp_attestation.rs (wire format)
