+++ core/src/lib.rs
@@ -27,11 +27,6 @@ pub mod ed25519_sigverifier;
 pub mod fetch_stage;
 pub mod forwarding_stage;
 pub mod gen_keys;
-pub mod mcp_attestation_service;
-pub mod mcp_bankless;
-pub mod mcp_consensus_broadcast;
-pub mod mcp_replay;
-pub mod mcp_voting;
 pub mod next_leader;
 pub mod optimistic_confirmation_verifier;
 pub mod repair;
