+++ ledger/src/lib.rs
@@ -27,8 +27,6 @@ pub mod leader_schedule_cache;
 pub mod leader_schedule_utils;
 pub mod mcp;
 pub mod mcp_attestation;
-pub mod mcp_schedule;
-pub mod mcp_schedule_cache;
 pub mod next_slots_iterator;
 pub mod rooted_slot_iterator;
 conditional_mod::conditional_vis_mod!(shred, feature="agave-unstable-api", pub,pub(crate));
