+++ svm/src/lib.rs
@@ -5,8 +5,6 @@
 
 pub mod account_loader;
 pub mod account_overrides;
-pub mod mcp_fee_payer;
-pub mod mcp_fee_replay;
 pub mod message_processor;
 pub mod nonce_info;
 pub mod program_loader;
