+++ turbine/src/lib.rs
@@ -4,12 +4,6 @@
 mod addr_cache;
 
 #[cfg(feature = "agave-unstable-api")]
-pub mod mcp_proposer;
-
-#[cfg(feature = "agave-unstable-api")]
-pub mod mcp_relay;
-
-#[cfg(feature = "agave-unstable-api")]
 pub mod broadcast_stage;
 
 #[cfg(feature = "agave-unstable-api")]
